requirements.txt
# Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database e Supabase
supabase==2.0.0
psycopg[binary]>=3.1.0
sqlalchemy==2.0.23

# Autenticazione e sicurezza
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# HTTP Client
httpx>=0.24.0,<0.25.0
requests==2.31.0

# Validazione dati
pydantic>=2.9.0
pydantic-settings>=2.5.0
email-validator>=2.0.0

# Utilità
python-dateutil==2.8.2

# Testing (opzionale per sviluppo)
pytest==7.4.3
pytest-asyncio==0.21.1

